Move bulk-edit-related API tests
